fix(simple-example): WARNING on OpenAPI schema generation, unify duplicate dataobjects (#63)
* avoid duplicate dataobjects
