+++ apps/examples/simple-example/src/model/__init__.py
@@ -65,3 +65,20 @@ class SomethingNotFound:
     """Item not found in datastore"""
     path: str
     id: str
+
+
+@dataobject
+@dataclass
+class ItemsInfo:
+    """
+    Items to read concurrently
+    """
+    item1_id: str
+    item2_id: str
+
+
+@dataobject(event_id='payload.id', event_ts='payload.status.ts')
+@dataclass
+class SomethingStored:
+    path: str
+    payload: Something
